Update property types and relationships in QPSettings
Changed data types from `int` to `long` for several properties in the `QPTemplateInstitutionDepartment`, `QPTemplateInstitutionDocument`, `QPTemplateStatusType`, and `UserQPTemplateDocument` classes to support larger values. Added `QPTemplateId` and `QPTemplateInstitutionId` properties in related classes to establish proper relationships. Included a new using directive for `DocumentFormat.OpenXml.Bibliography` for potential document formatting features.

+++ SKCE.Valuation.API/Examination.Models/DbModels/QPSettings/QPTemplateInstitutionDepartment.cs
@@ -14,10 +14,10 @@ namespace SKCE.Examination.Models.DbModels.QPSettings
     {
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int QPTemplateInstitutionDepartmentId { get; set; }
+        public long QPTemplateInstitutionDepartmentId { get; set; }
         
         [ForeignKey("QPTemplateInstitution")]
-        public int QPTemplateInstitutionId { get; set; }
+        public long QPTemplateInstitutionId { get; set; }
         public long DepartmentId { get; set; }
         public long StudentCount { get; set; }
     }
